refactoring a lot of About Page
