+++ frontend/package.json
@@ -25,6 +25,7 @@
 		"react-spinners": "^0.2.6",
 		"react-twitter-widgets": "^1.7.1",
 		"recharts": "^1.0.0-beta.10",
+		"request-promise": "^4.2.2",
 		"requests": "^0.2.2",
 		"slick-carousel": "^1.8.1"
 	},
