+++ frontend/src/js/About/memberData.js
@@ -1,0 +1,34 @@
+export function getMembers(){
+	let sweMembers = {}
+	sweMembers['MTirtowidjojo'] = ['Michael Tirtowidjojo', 0, 0, 0,
+	  'Michael is a third-year CS student who trains in Taekwondo and ' +
+	  'enjoys reading World War II stories.', 'Michael.png',
+	  'Parties front-end and chief of chief naming', 'Mojo Jojo', 'TX', '14']
+	sweMembers['copperstick6'] = ['William Han', 0, 0, 68,
+	  'William is a sophomore CS student who enjoys the subtle art of ' +
+	  'memes and hackathons.', 'William.jpg',
+	  'Districts front-end and backend-unit tests ' +
+	  'deployment', 'Bill', 'TX', '12']
+	sweMembers['raulcodes'] = ['Raul Camacho', 0, 0, 19,
+	  'Raul is a senior CS major who will be graduating this semester. ' +
+	  'Also he\'s tall. Probably too tall.', 'Raul.png',
+	  'Front-end, css, front-end unittests and acceptance tests',
+	  'Camacho Style Sheets', 'TX', '12']
+	sweMembers['minwoo0jo'] = ['Minwoo Jo', 0, 0, 0,
+	  'Minwoo is a fourth year student currently pursuing a BSA in CS. He ' +
+	  'enjoys studying foreign languages and competing in video game ' +
+	  'tournaments in his free time.', 'Minwoo.jpg',
+	  'District front-end and back-end, API designer', 'MJ', 'TX', '24']
+	sweMembers['bzsinger'] = ['Benjamin Singer', 0, 0, 0,
+	  'Benny is a third-year CS student who enjoys iOS development, ' +
+	  'reading, and following current events.', 'Benny.jpg',
+	  'Representatives front-end, District back-end, Documentation', 'Benny', 'TX', '32']
+	sweMembers['palakhirpara'] = ['Palak Hirpara', 0, 0, 34,
+
+	  'Palak is a senior who will be graduating this semester with BSCS ' +
+	  'and likes watching cricket.', 'Palak.png',
+	  'Representatives front-end and back-end, Districts back-end, back-end unittests',
+	  'The Decider', 'TX', '32']
+	  return sweMembers;
+
+}
