+++ frontend/src/js/Router.js
@@ -3,7 +3,7 @@ import React, {Component} from 'react'
 import {Route, Switch} from 'react-router-dom'
 /* eslint-enable no-unused-vars */
 
-import About from './About/About'
+import About from './About/About.jsx'
 import Alma from './Representatives/Alma.js'
 import Districts from './Districts/Districts'
 import DistrictsDetails from './Districts/DistrictDetails'
